Regroup the 2 classes
